Add createRegistration to useRegistrationActions hook refactor create/edit registration pages
